# backend-api-handler/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Travel Agency API",
    "description": "Sprint 2 — Auth, Tours, Bookings, and Tour Booking Lifecycle Management. Includes JWT-based authentication, tour browsing, booking lifecycle (BOOKED → CONFIRMED → STARTED → FINISHED / CANCELED), document upload & verification, and travel-agent dedicated endpoints.",
    "contact": {
      "name": "Travel Agency Team",
      "email": "support@travelagency.com"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "security": [
    {
      "Bearer": []
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/sign-up": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new user account",
        "operationId": "signUp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpResponseDTO"
                }
              }
            }
          },
          "409": {
            "description": "Email already exists",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sign-in": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Sign in with email and password",
        "operationId": "signIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too many requests",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponseDTO"
                }
              }
            }
          },
          "200": {
            "description": "Authentication successful",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Refresh access token using refresh token cookie",
        "operationId": "refresh",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired refresh token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout and invalidate refresh token",
        "operationId": "logout",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SignUpRequestDTO": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "lastName": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 16,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "SignUpResponseDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "SignInRequestDTO": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 16,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "SignInResponseDTO": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste your JWT access token here",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}